add results for 80:20 etc ratio, weekends-weekdays, & 1 month train - 1 week test
