+++ Occupancy/Results/.~lock.month_train_week_test.ods#
@@ -1,0 +1,1 @@
+,NOBODY/NVT,Nobody,13.09.2016 18:09,file:///C:/Users/NVT/AppData/Roaming/LibreOffice/4;
