+++ Occupancy/result_train_month_jul.csv
@@ -1,0 +1,45 @@
+
+r2,1,300,0.520833333333
+r2,1,600,0.557291666667
+r2,1,900,0.614583333333
+r2,1,1800,0.390625
+r2,1,3600,0.229166666667
+r2,2,300,0.500868055556
+r2,2,600,0.461805555556
+r2,2,900,0.317708333333
+r2,2,1800,0.28125
+r2,2,3600,0.208333333333
+r2,3,300,0.340277777778
+r2,3,600,0.295138888889
+r2,3,900,0.2421875
+r2,3,1800,0.265625
+r2,3,3600,0.208333333333
+r2,4,300,0.279513888889
+r2,4,600,0.232638888889
+r2,4,900,0.231770833333
+r2,4,1800,0.255208333333
+r2,4,3600,0.208333333333
+r2,5,300,0.295138888889
+r2,5,600,0.279513888889
+r2,5,900,0.244791666667
+r2,5,1800,0.260416666667
+r2,5,3600,0.208333333333
+r2,10,300,0.327256944444
+r2,10,600,0.3125
+r2,10,900,0.263020833333
+r2,10,1800,0.276041666667
+r2,10,3600,0.208333333333
+r2,30,300,0.327256944444
+r2,30,600,0.286458333333
+r2,30,900,0.25
+r2,30,1800,0.270833333333
+r2,30,3600,0.208333333333
+r2,60,300,0.335069444444
+r2,60,600,0.288194444444
+r2,60,900,0.276041666667
+r2,60,1800,0.270833333333
+r2,60,3600,0.208333333333
+r2,300,1800,0.25
+r2,300,3600,0.208333333333
+r2,600,3600,0.208333333333
+r2,900,3600,0.208333333333
