+++ Occupancy/result_weekends_1.csv
@@ -1,0 +1,216 @@
+
+r2,0.1,1,300,0.797970984602
+r2,0.1,1,600,0.582977534448
+r2,0.1,1,900,0.62266313933
+r2,0.1,1,1800,0.675925925926
+r2,0.1,1,3600,0.71875
+r2,0.1,2,300,0.810456697299
+r2,0.1,2,600,0.681685906686
+r2,0.1,2,900,0.701366843034
+r2,0.1,2,1800,0.675925925926
+r2,0.1,2,3600,0.71875
+r2,0.1,3,300,0.722156738648
+r2,0.1,3,600,0.681685906686
+r2,0.1,3,900,0.701366843034
+r2,0.1,3,1800,0.675925925926
+r2,0.1,3,3600,0.71875
+r2,0.1,4,300,0.722156738648
+r2,0.1,4,600,0.681685906686
+r2,0.1,4,900,0.701366843034
+r2,0.1,4,1800,0.675925925926
+r2,0.1,4,3600,0.71875
+r2,0.1,5,300,0.722156738648
+r2,0.1,5,600,0.681685906686
+r2,0.1,5,900,0.701366843034
+r2,0.1,5,1800,0.675925925926
+r2,0.1,5,3600,0.71875
+r2,0.1,10,300,0.722156738648
+r2,0.1,10,600,0.681685906686
+r2,0.1,10,900,0.701366843034
+r2,0.1,10,1800,0.675925925926
+r2,0.1,10,3600,0.71875
+r2,0.1,30,300,0.722156738648
+r2,0.1,30,600,0.681685906686
+r2,0.1,30,900,0.701366843034
+r2,0.1,30,1800,0.675925925926
+r2,0.1,30,3600,0.71875
+r2,0.1,60,300,0.722156738648
+r2,0.1,60,600,0.681685906686
+r2,0.1,60,900,0.701366843034
+r2,0.1,60,1800,0.675925925926
+r2,0.1,60,3600,0.71875
+r2,0.1,300,1800,0.675925925926
+r2,0.1,300,3600,0.71875
+r2,0.1,600,3600,0.71875
+r2,0.1,900,3600,0.71875
+r2,0.2,1,300,0.991915056621
+r2,0.2,1,600,0.32175347132
+r2,0.2,1,900,0.590739933871
+r2,0.2,1,1800,0.679783950617
+r2,0.2,1,3600,0.765740740741
+r2,0.2,2,300,0.844208453941
+r2,0.2,2,600,0.694147343004
+r2,0.2,2,900,0.711785646557
+r2,0.2,2,1800,0.679783950617
+r2,0.2,2,3600,0.765740740741
+r2,0.2,3,300,0.718732703465
+r2,0.2,3,600,0.696562801941
+r2,0.2,3,900,0.711785646557
+r2,0.2,3,1800,0.679783950617
+r2,0.2,3,3600,0.765740740741
+r2,0.2,4,300,0.718732703465
+r2,0.2,4,600,0.696562801941
+r2,0.2,4,900,0.711785646557
+r2,0.2,4,1800,0.679783950617
+r2,0.2,4,3600,0.765740740741
+r2,0.2,5,300,0.718732703465
+r2,0.2,5,600,0.696562801941
+r2,0.2,5,900,0.711785646557
+r2,0.2,5,1800,0.679783950617
+r2,0.2,5,3600,0.765740740741
+r2,0.2,10,300,0.718732703465
+r2,0.2,10,600,0.696562801941
+r2,0.2,10,900,0.711785646557
+r2,0.2,10,1800,0.679783950617
+r2,0.2,10,3600,0.765740740741
+r2,0.2,30,300,0.718732703465
+r2,0.2,30,600,0.696562801941
+r2,0.2,30,900,0.711785646557
+r2,0.2,30,1800,0.679783950617
+r2,0.2,30,3600,0.765740740741
+r2,0.2,60,300,0.718732703465
+r2,0.2,60,600,0.696562801941
+r2,0.2,60,900,0.711785646557
+r2,0.2,60,1800,0.679783950617
+r2,0.2,60,3600,0.765740740741
+r2,0.2,300,1800,0.679783950617
+r2,0.2,300,3600,0.765740740741
+r2,0.2,600,3600,0.765740740741
+r2,0.2,900,3600,0.765740740741
+r2,0.3,1,300,0.514261887672
+r2,0.3,1,600,0.461610572191
+r2,0.3,1,900,0.631214355724
+r2,0.3,1,1800,0.68284030784
+r2,0.3,1,3600,0.750132275132
+r2,0.3,2,300,0.794955192002
+r2,0.3,2,600,0.706132753858
+r2,0.3,2,900,0.702015236942
+r2,0.3,2,1800,0.68284030784
+r2,0.3,2,3600,0.750132275132
+r2,0.3,3,300,0.726684371421
+r2,0.3,3,600,0.710174494011
+r2,0.3,3,900,0.704540489467
+r2,0.3,3,1800,0.68284030784
+r2,0.3,3,3600,0.750132275132
+r2,0.3,4,300,0.72855297264
+r2,0.3,4,600,0.710174494011
+r2,0.3,4,900,0.704540489467
+r2,0.3,4,1800,0.68284030784
+r2,0.3,4,3600,0.750132275132
+r2,0.3,5,300,0.72855297264
+r2,0.3,5,600,0.710174494011
+r2,0.3,5,900,0.704540489467
+r2,0.3,5,1800,0.68284030784
+r2,0.3,5,3600,0.750132275132
+r2,0.3,10,300,0.72855297264
+r2,0.3,10,600,0.710174494011
+r2,0.3,10,900,0.704540489467
+r2,0.3,10,1800,0.68284030784
+r2,0.3,10,3600,0.750132275132
+r2,0.3,30,300,0.72855297264
+r2,0.3,30,600,0.710174494011
+r2,0.3,30,900,0.704540489467
+r2,0.3,30,1800,0.68284030784
+r2,0.3,30,3600,0.750132275132
+r2,0.3,60,300,0.72855297264
+r2,0.3,60,600,0.710174494011
+r2,0.3,60,900,0.704540489467
+r2,0.3,60,1800,0.68284030784
+r2,0.3,60,3600,0.750132275132
+r2,0.3,300,1800,0.68284030784
+r2,0.3,300,3600,0.750132275132
+r2,0.3,600,3600,0.750132275132
+r2,0.3,900,3600,0.750132275132
+r2,0.4,1,300,0.362250815593
+r2,0.4,1,600,0.486961939572
+r2,0.4,1,900,0.690068108775
+r2,0.4,1,1800,0.683494038641
+r2,0.4,1,3600,0.741115520282
+r2,0.4,2,300,0.716409067343
+r2,0.4,2,600,0.707527963816
+r2,0.4,2,900,0.712809811517
+r2,0.4,2,1800,0.683494038641
+r2,0.4,2,3600,0.741115520282
+r2,0.4,3,300,0.719822868048
+r2,0.4,3,600,0.707527963816
+r2,0.4,3,900,0.712809811517
+r2,0.4,3,1800,0.683494038641
+r2,0.4,3,3600,0.741115520282
+r2,0.4,4,300,0.719822868048
+r2,0.4,4,600,0.707527963816
+r2,0.4,4,900,0.712809811517
+r2,0.4,4,1800,0.683494038641
+r2,0.4,4,3600,0.741115520282
+r2,0.4,5,300,0.719822868048
+r2,0.4,5,600,0.707527963816
+r2,0.4,5,900,0.712809811517
+r2,0.4,5,1800,0.683494038641
+r2,0.4,5,3600,0.741115520282
+r2,0.4,10,300,0.719822868048
+r2,0.4,10,600,0.707527963816
+r2,0.4,10,900,0.712809811517
+r2,0.4,10,1800,0.683494038641
+r2,0.4,10,3600,0.741115520282
+r2,0.4,30,300,0.719822868048
+r2,0.4,30,600,0.707527963816
+r2,0.4,30,900,0.712809811517
+r2,0.4,30,1800,0.683494038641
+r2,0.4,30,3600,0.741115520282
+r2,0.4,60,300,0.719822868048
+r2,0.4,60,600,0.707527963816
+r2,0.4,60,900,0.712809811517
+r2,0.4,60,1800,0.683494038641
+r2,0.4,60,3600,0.741115520282
+r2,0.4,300,1800,0.683494038641
+r2,0.4,300,3600,0.741115520282
+r2,0.4,600,3600,0.741115520282
+r2,0.4,900,3600,0.741115520282
+r2,0.5,1,300,0.813241862432
+r2,0.5,1,600,0.933793929218
+r2,0.5,1,900,0.709002585975
+r2,0.5,1,1800,0.708439273145
+r2,0.5,1,3600,0.735119047619
+r2,0.5,2,300,0.729261447673
+r2,0.5,2,600,0.712672771318
+r2,0.5,2,900,0.714893661291
+r2,0.5,2,1800,0.708439273145
+r2,0.5,2,3600,0.735119047619
+r2,0.5,3,300,0.715198612412
+r2,0.5,3,600,0.712672771318
+r2,0.5,3,900,0.714893661291
+r2,0.5,3,1800,0.708439273145
+r2,0.5,3,3600,0.735119047619
+r2,0.5,4,300,0.715198612412
+r2,0.5,4,600,0.712672771318
+r2,0.5,4,900,0.714893661291
+r2,0.5,4,1800,0.708439273145
+r2,0.5,4,3600,0.735119047619
+r2,0.5,5,300,0.715198612412
+r2,0.5,5,600,0.712672771318
+r2,0.5,5,900,0.714893661291
+r2,0.5,5,1800,0.708439273145
+r2,0.5,5,3600,0.735119047619
+r2,0.5,10,300,0.715198612412
+r2,0.5,10,600,0.712672771318
+r2,0.5,10,900,0.714893661291
+r2,0.5,10,1800,0.708439273145
+r2,0.5,10,3600,0.735119047619
+r2,0.5,30,300,0.715198612412
+r2,0.5,30,600,0.712672771318
+r2,0.5,30,900,0.714893661291
+r2,0.5,30,1800,0.708439273145
+r2,0.5,30,3600,0.735119047619
+r2,0.5,60,300,0.715198612412
+r2,0.5,60,600,0.712672771318
+r2,0.5,60,900,0.714893661291
+r2,0.5,60,1800,0.708439273145
